refactor!: extract thumbnail related things to thumbnail.c
breaking change: image_thumbnail() takes arguments
